Split memory usage collection and reporting
It will be useful in some places of the code to measure memory usage but
not necessarily to display it as a normal "report" in the logs. Thus,
splitting metrics collection from reporting will help us reuse the
collection bits in other places of the code.

While doing this I also took the chance of simplifying how these memory
collection metrics are performed. For starters, we don't need to know
about that many fields, as we're usually only interested on VM/RSS
curernt/peak usage. Additionally we can use a single file as the source
of the information, which should help speeding up this data collection
as well.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>

Use a simpler function names in memory reports

Pretty names are unnecessarily long: they don't really give any useful
information, specially because we already have the location in the file
where the report is generated

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>

Simplify memory usage reporting calls

Memory reporting doesn't really need to use the Options object anymore
(it was previously used to keep track of peak memory usage, which the
kernel already does for us); additionally we already have the file/line
location showing up on the log statement when we generate a memory
report, so there's no need to have the information twice in the same
line.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>

Remove memuse_* fields from Options class

These were not used in reality, or were previously not used correctly,
so there's no need to keep them around. If we need some memory tracking
functionality in the future, we can reinstate them in a different form.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>

Finish VR graciously on uncaught exceptions

VR's main function didn't have a high-level try/catch block to handle
uncaught exceptions rising from the code, leaving the behavior of the
program to the fate of whatever default the C++ runtime library
provides.

This commit adds such a block. When an uncaught exception is received,
an error is logged, and then either an MPI_Abort is issued in
COMM_WORLD, or a simpler exit is invoked on non-MPI builds.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>

Log memory requirements for data transfers

This might help us understand what exactly is causing #118 (if logs are
correctly flushed and don't get too scrambled).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>